+++ test/msc3916_misc_client_endpoints_suite_test.go
@@ -57,6 +57,7 @@ func (s *MSC3916MiscClientEndpointsSuite) TestPreviewUrlRequiresAuth() {
 		AccessToken:     "", // no auth on this client
 		UserId:          "", // no auth on this client
 	}
+	clientGuest := s.deps.Homeservers[0].GuestUsers[0].WithCsUrl(s.deps.Machines[0].HttpUrl)
 
 	qs := url.Values{
 		"url": []string{s.htmlPage.PublicUrl},
@@ -64,6 +65,10 @@ func (s *MSC3916MiscClientEndpointsSuite) TestPreviewUrlRequiresAuth() {
 	raw, err := client2.DoRaw("GET", "/_matrix/client/v1/media/preview_url", qs, "", nil)
 	assert.NoError(t, err)
 	assert.Equal(t, http.StatusUnauthorized, raw.StatusCode)
+
+	raw, err = clientGuest.DoRaw("GET", "/_matrix/client/v1/media/preview_url", qs, "", nil)
+	assert.NoError(t, err)
+	assert.Equal(t, http.StatusForbidden, raw.StatusCode)
 
 	raw, err = client1.DoRaw("GET", "/_matrix/client/v1/media/preview_url", qs, "", nil)
 	assert.NoError(t, err)
@@ -80,10 +85,15 @@ func (s *MSC3916MiscClientEndpointsSuite) TestConfigRequiresAuth() {
 		AccessToken:     "", // no auth on this client
 		UserId:          "", // no auth on this client
 	}
+	clientGuest := s.deps.Homeservers[0].GuestUsers[0].WithCsUrl(s.deps.Machines[0].HttpUrl)
 
 	raw, err := client2.DoRaw("GET", "/_matrix/client/v1/media/config", nil, "", nil)
 	assert.NoError(t, err)
 	assert.Equal(t, http.StatusUnauthorized, raw.StatusCode)
+
+	raw, err = clientGuest.DoRaw("GET", "/_matrix/client/v1/media/config", nil, "", nil)
+	assert.NoError(t, err)
+	assert.Equal(t, http.StatusForbidden, raw.StatusCode)
 
 	raw, err = client1.DoRaw("GET", "/_matrix/client/v1/media/config", nil, "", nil)
 	assert.NoError(t, err)