+++ test/test_internals/deps.go
@@ -98,7 +98,7 @@ func MakeTestDeps() (*ContainerDeps, error) {
 
 	// Start postgresql database
 	pgContainer, err := postgres.RunContainer(ctx,
-		testcontainers.WithImage("docker.io/library/postgres:14"),
+		testcontainers.WithImage("docker.io/library/postgres:16"),
 		postgres.WithDatabase("mmr"),
 		postgres.WithUsername("postgres"),
 		postgres.WithPassword("test1234"),
@@ -115,10 +115,14 @@ func MakeTestDeps() (*ContainerDeps, error) {
 	}
 	// we can hardcode the port and most of the connection details because we're behind the docker network here
 	pgConnStr := fmt.Sprintf("host=%s port=5432 user=postgres password=test1234 dbname=mmr sslmode=disable", pgHost)
+	// the external connection string is a bit harder, because testcontainers wants to use `localhost` as a hostname,
+	// which prevents us from using the ConnectionString() function. We instead build the connection string manually
+	//pgExtPort, err := pgContainer.MappedPort(ctx, "5432/tcp")
 	extPgConnStr, err := pgContainer.ConnectionString(ctx, "sslmode=disable")
 	if err != nil {
 		return nil, err
 	}
+	//extPgConnStr := fmt.Sprintf("host=%s port=%d user=postgres password=test1234 dbname=mmr sslmode=disable", testcontainers.HostInternal, pgExtPort.Int())
 
 	// Start a redis container
 	cwd, err := os.Getwd()